feat: migrate backend hardware events to ui_post_notification (plan v7 phase 3)
Replace 16 GuiApiEmitSignal calls in backend with ui_post_notification:

- background_task.c (7): USB connection/disconnect, battery state, SD card
- low_power.c (3): SD card after wakeup, USB connection after wakeup
- screen_manager.c (2): status bar refresh, close message box
- power_manager.c (1): power option button
- usb_task.c (1): USB state change
- fingerprint_process.c (2): lock screen fingerprint pass/fail

Frontend change:
- gui_init_view.c: SIG_INIT_BATTERY reads uint32_t instead of uint16_t
  (matches ui_post_notification's uint32_t value parameter)

+++ src/ui/gui_views/gui_init_view.c
@@ -76,7 +76,7 @@ int32_t GUI_InitViewEventProcess(void *self, uint16_t usEvent, void *param, uint
 {
     static uint8_t walletNum;
     static uint16_t lockParam = SIG_LOCK_VIEW_VERIFY_PIN;
-    uint16_t battState;
+    uint32_t battState;
     uint32_t rcvValue;
 
     switch (usEvent) {
@@ -101,7 +101,7 @@ int32_t GUI_InitViewEventProcess(void *self, uint16_t usEvent, void *param, uint
         }
         break;
     case SIG_INIT_BATTERY:
-        battState = *(uint16_t *)param;
+        battState = *(uint32_t *)param;
         //printf("rcv battState=0x%04X\r\n", battState);
         GuiStatusBarSetBattery(battState & 0xFF, (battState & 0x8000) != 0);
         break;
